drm/i915/psr: Move plane sel fetch configuration into plane source files
commit b1f5279b5981f9ed851163ee661692f42397982f upstream.

Currently selective fetch configuration for planes is implemented in psr
code. More suitable place for this code is where everything else is
configured for planes -> move it into skl_universal_plane.c and
intel_cursor.c. This also allows us to drop hooks for cursor handling.

v3: Checkpatch warnings fixed
v2: Removed setting sel_fetch_area->y1/y2 as -1

deepin-Intel-SIG: commit b1f5279b5981 drm/i915/psr: Move plane sel fetch configuration into plane source files.
MTL Graphics backporting - 2024Q2

Link: https://patchwork.freedesktop.org/patch/msgid/20231120082606.3156488-2-jouni.hogander@intel.com
[ Quanxian Wang: amend commit log ]
